Excepted functionality defined with unit tests.

+++ Domain/Interfaces/IQualityEvaluator.cs
@@ -4,6 +4,6 @@ namespace Domain.Interfaces
 {
     public interface IQualityEvaluator<T> where T : ItemBase
     {
-        decimal EvaluateQualityOf(T item);
+        int EvaluateQualityOf(T item);
     }
 }
